\name{HyperLink}
\alias{HyperLink}
\docType{data}
\title{
A Adjacent Matrix of Wikipedia, which means the relation of hyper-link
}
\description{
This matrix has *** rows and *** columns (documents).
}
\usage{data("HyperLink")}
\format{
  The format is:
 int [1:6000, 1:6000] 0 0 0 0 1 0 0 0 0 1 ...
}
\source{
Pokémon_Go
https://en.wikipedia.org/wiki/Pokémon_Go
}
\examples{
data(HyperLink)
}
\keyword{datasets}
